test: 예약 가능 시간 및 상태 조회 테스트
